backend/requirements.txt
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
uvicorn==0.37.0
groq>=0.4.0
python-dotenv>=1.0.0
sqlalchemy >= 2.0.44
psycopg2-binary >= 2.9.12